move reactions listener to app (#349)
