+++ src/lecturer/index.tsx
@@ -1,15 +1,23 @@
+import "./index.css";
+
 import React from "react";
 import ReactDOM from "react-dom";
-import "./index.css";
+import { BrowserRouter as Router } from "react-router-dom";
+
 import App from "./components/app/App";
-import * as serviceWorkerRegistration from "./serviceWorkerRegistration";
 import reportWebVitals from "./reportWebVitals";
 import { SocketService } from "./services/SocketService";
+import Store from "./services/StoreService";
+import * as serviceWorkerRegistration from "./serviceWorkerRegistration";
 
 ReactDOM.render(
     <React.StrictMode>
         <SocketService>
-            <App />
+            <Store>
+                <Router>
+                    <App />
+                </Router>
+            </Store>
         </SocketService>
     </React.StrictMode>,
     document.getElementById("root")
